+++ cmd/carbonapi/http/helper.go
@@ -118,11 +118,11 @@ var knownFormats = map[string]responseFormat{
 	"protobuf3":       protoV2Format,
 	"carbonapi_v2_pb": protoV2Format,
 	"carbonapi_v3_pb": protoV3Format,
-	"png": pngFormat,
-	"csv": csvFormat,
-	"raw": rawFormat,
-	"svg": svgFormat,
-	"completer": completerFormat,
+	"png":             pngFormat,
+	"csv":             csvFormat,
+	"raw":             rawFormat,
+	"svg":             svgFormat,
+	"completer":       completerFormat,
 }
 
 const (